ref: Run a query in the context of a connection
This commit adds a new method to the Connection class called run which
takes in a RethinkDB query object and runs it using its internal
RethinkDB connection.  This simplifies the code that needs to be written
for database interactions as individual tables classes don't need to
know about the internal workings of the Connection class.