feat: Insert new artists into the database
This commit hooks up the artist POST route to the database so that new
artists will be saved.  In the route itself we need to delete the ID
field since the API model is going to be shared between all artist
routes and the POST request is intended to always create a new item.