fix: Do not attempt to insert an empty artist list
This commit fixes a bug in the artist insert function.  If the function
was called with an empty list of artists to create it would execute the
database query but fail when trying to get the "generated_keys" to
return.  When there is no data to insert we don't even need to get that
far and can short circuit with a quick return.