feat: Deal with null values in the track remix
This commit adds support for null values in the track remix field.  This
is done by using the `allow_null` parameter of the nested field class
which causes the `marshal` function to return null if all the nested
fields are null.  Additionally, the verbose tracklist request needs to
be rewritten to branch on the state of the remix field and prevent a
query error when joining on the remix artist.