fix: Dedupe different versions of the same track
This commit fixes a bug with the track deduplication code where two
versions of the same track in a single bulk import would get marked as
the same since they have the same name.  This is now fixed by performing
a hash on the JSON model and using that hash code as the key.

fixes #16