fix: Map tracks by hash not by name
This commit fixes a bug with the bulk import code where the track map to
get an ID was keyed on the name which is not a unique identifier.  This
is fixed by using the Python hash function on an object which is
marshalled from the artist model with the ID removed.  Essentially this
means that we are getting a unique key based on track name, artist name,
remix name, and remix artist name.