feat: Link alternate versions to the original
This commit adds code to the bulk import route that loops over the
original tracks related to the import and updates the versions array on
each one to contain all of the related track IDs.  This will be useful
for navigating through the database.  At the moment this is really slow.

closes #18