perf: Index tracks by the originalId field
This commit improves the performance of the bulk import code by adding
an index on the originalId field of the track table.  This can then be
used to search for tracks by their original track ID in bulk.
Performance is also improved by modifying the code to operate in bulk
instead of one track at a time.