ref: Standardize the artist nested schema
This commit refactors the way that artists are stored when nested in
other schemas.  Instead of storing the ID directly, we now store it as
an `id` property inside of an `artist` object.  While this adds slight
complexity at write time, it creates a simpler experience when
retrieving data with the verbose flag.