ref: Store tracks as nested objects in tracklists
This commit refactors how tracks are stored within tracklists.  Instead
of storing the ID at the top level next to the index we now store a
nested track object.  This makes for a simpler experience when
retrieving data and should also allow us to more simply map to the bulk
import requests.