fix: Use the new marshmallow models in imports
This commit fixes all the bulk import code to use the new marshmallow
models and schemas.  It also adds a function to the Track class to get
a unique hash for a track based on the unique attributes other than the
track's ID which can be used to simplify the import utilities.